feat: interceptor 기능 instance에 추가
