+++ src/api/instance.ts
@@ -1,3 +1,4 @@
+import { clearTokens, getAccessToken, getRefreshToken } from '@/utils/auth';
 import axios from 'axios';
 
 export const open_api_axiosInstance = axios.create({
@@ -12,3 +13,28 @@ export const llm_server_api_axiosInstance = axios.create({
 export const spring_server_api_axiosInstance = axios.create({
   baseURL: process.env.NEXT_PUBLIC_SPRING_SERVER_API_ADDRESS,
 });
+
+spring_server_api_axiosInstance.interceptors.request.use((config) => {
+  if (typeof document !== 'undefined') {
+    const accessToken = getAccessToken();
+    const refreshToken = getRefreshToken();
+    if (accessToken) {
+      config.headers['authorization'] = accessToken;
+    }
+    if (refreshToken) {
+      config.headers['authorization-refresh'] = refreshToken;
+    }
+  }
+  return config;
+});
+
+spring_server_api_axiosInstance.interceptors.response.use(
+  (response) => response,
+  (error) => {
+    if (error.response?.status === 401 && typeof window !== 'undefined') {
+      clearTokens();
+      window.location.href = '/login';
+    }
+    return Promise.reject(error);
+  },
+);
